add a new endpoint that will periodically polled to check number is still connected (#21)
* I wanna add a new endpoint that will periodically polled to check that the system is operational.

Its a good time to start a refactor of our current other 2 tasks that don't work with the server.
To simplify Ops, I wanna call them via an endpoint instead of imlementing the logic in a separated script. This will be a task down the road.

So this is partial, first create the endpoint, then add a cron that checks it, error if the cron fails ( ops part with our render ops implementation )

* update cursor rules

* lint fix

* Add script to check status.
Implement endpoint logic